requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
abseil-cpp=20230802.0=h6a678d5_2
absl-py=2.1.0=py311h06a4308_0
aiohttp=3.9.3=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
attrs=23.2.0
blas=1.0=openblas
blinker=1.6.2=py311h06a4308_0
brotli-python=1.0.9=py311h6a678d5_8
bzip2=1.0.8=hd590300_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=py311h06a4308_0
certifi=2024.2.2=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py311h06a4308_0
contextlib2=21.6.0
cpuonly=2.0=0
cryptography=42.0.5=py311hdda0065_0
filelock=3.13.1=py311h06a4308_0
frozenlist=1.4.0=py311h5eee18b_0
fsspec=2024.3.1=py311h06a4308_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
google-auth=2.22.0=py311h06a4308_0
google-auth-oauthlib=0.5.2=py311h06a4308_0
grpc-cpp=1.48.2=he1ff14a_4
grpcio=1.48.2=py311he1ff14a_4
gtest=1.14.0=hdb19cb5_0
idna=3.7=py311h06a4308_0
jinja2=3.1.3=py311h06a4308_0
ld_impl_linux-64=2.40=h55db66e_0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h77fa898_6
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h77fa898_6
libllvm14=14.0.6=hdb19cb5_3
libnsl=2.0.1=hd590300_0
libopenblas=0.3.21=h043d6bf_0
libprotobuf=3.20.3=he621ea3_0
libsqlite=3.45.3=h2797004_0
libstdcxx-ng=13.2.0=hc0a3c3a_6
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvmlite=0.42.0=py311h6a678d5_0
markdown=3.4.1=py311h06a4308_0
markupsafe=2.1.3=py311h5eee18b_0
ml-collections=0.1.1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.4=py311h5eee18b_0
ncurses=6.4.20240210=h59595ed_0
networkx=3.1=py311h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.59.1=py311h96b013e_0
numpy=1.26.4=py311h24aa872_0
numpy-base=1.26.4=py311hbfb1bba_0
oauthlib=3.2.2=py311h06a4308_0
openssl=3.3.0=hd590300_0
pip=24.0=pyhd8ed1ab_0
protobuf=3.20.3=py311h6a678d5_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py311h06a4308_0
pyopenssl=24.0.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.9=hb806964_0_cpython
python_abi=3.11=4_cp311
pytorch=2.2.0=cpu_py311h53e38e9_0
pytorch-mutex=1.0=cpu
pyyaml=6.0.1
re2=2022.04.01=h295c915_0
readline=8.2=h8228510_1
requests=2.31.0=py311h06a4308_1
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.13.0
setuptools=69.5.1=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
sympy=1.12=py311h06a4308_0
tensorboard=2.12.1=py311h06a4308_0
tensorboard-data-server=0.7.0=py311h52d8a92_0
tensorboard-plugin-wit=1.6.0=py_0
tk=8.6.13=noxft_h4845f30_101
typing_extensions=4.9.0=py311h06a4308_1
tzdata=2024a=h0c530f3_0
urllib3=1.26.18=py311h06a4308_0
werkzeug=2.3.8=py311h06a4308_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h166bdaf_0
yarl=1.9.3=py311h5eee18b_0
zlib=1.2.13=hd590300_5
